Edit functionality of Project title is complete
Edit project title function implemented
